Add middleware into route

+++ backend_api_web_nghe_nhac/routes/api.php
@@ -25,27 +25,25 @@ Route::group([
 ], function ($router) {
     Route::post('/login', [AuthController::class, 'login']);
     Route::post('/register', [AuthController::class, 'register']);
-    Route::post('/logout', [AuthController::class, 'logout']);
-    Route::post('/refresh', [AuthController::class, 'refresh']);
-    Route::get('/user-profile', [AuthController::class, 'userProfile']);
-    Route::post('/change-password', [AuthController::class, 'changePassword']);
+    Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth');
+    Route::post('/refresh', [AuthController::class, 'refresh'])->middleware('auth');
+    Route::get('/user-profile', [AuthController::class, 'userProfile'])->middleware('auth');
+    Route::post('/change-password', [AuthController::class, 'changePassword'])->middleware('auth');
 });
 
 Route::get('/songs', [SongController::class,'index'])->name('songs.all');
 Route::get('/songs/{songId}', [SongController::class,'show'])->name('songs.show');
-Route::post('/songs', [SongController::class,'store'])->name('songs.store');
-Route::put('/songs/{songId}', [SongController::class,'update'])->name('songs.update');
-Route::delete('/songs/{songId}', [SongController::class,'destroy'])->name('songs.destroy');
+Route::post('/songs', [SongController::class,'store'])->name('songs.store')->middleware('auth');
+Route::put('/songs/{songId}', [SongController::class,'update'])->name('songs.update')->middleware('auth');
+Route::delete('/songs/{songId}', [SongController::class,'destroy'])->name('songs.destroy')->middleware('auth');
 Route::get('/latest/songs/',[SongController::class,'newSong'])->name('songs.latest');
 Route::get('/{id_singer}/songs', [SongController::class,'songsOfSinger'])->name('songs.singer');
 
 
 Route::get('/users', [UserController::class,'index'])->name('users.all');
 Route::get('/users/{userId}', [UserController::class,'show'])->name('users.show');
-
-
 Route::get('/singers', [SingerController::class,'index'])->name('singers.all');
 Route::get('/singers/{singerId}', [SingerController::class,'show'])->name('singers.show');
-Route::post('/singers', [SingerController::class,'store'])->name('singers.store');
-Route::put('/singers/{singerId}', [SingerController::class,'update'])->name('singers.update');
+Route::post('/singers', [SingerController::class,'store'])->name('singers.store')->middleware('auth');
+Route::put('/singers/{singerId}', [SingerController::class,'update'])->name('singers.update')->middleware('auth');
 Route::delete('/singers/{singerId}', [SingerController::class,'destroy'])->name('singers.destroy');
